extract out ByteArrayInputStream from VersionServiceTest
